+++ packages/eyes-selenium/test/util/version-alias.js
@@ -1,7 +1,10 @@
 const alias = require('module-alias')
 
-if (!process.env.SELENIUM_MAJOR_VERSION) {
-  process.env.SELENIUM_MAJOR_VERSION = '4'
+if (!process.env.APPLITOOLS_SELENIUM_MAJOR_VERSION) {
+  process.env.APPLITOOLS_SELENIUM_MAJOR_VERSION = '4'
 }
 
-alias.addAlias('selenium-webdriver', `selenium-webdriver-${process.env.SELENIUM_MAJOR_VERSION}`)
+alias.addAlias(
+  'selenium-webdriver',
+  `selenium-webdriver-${process.env.APPLITOOLS_SELENIUM_MAJOR_VERSION}`,
+)