+++ packages/sdk-coverage-tests/test/report/index.spec.js
@@ -60,7 +60,7 @@ describe('Report', () => {
   })
   it('should return the report test name', () => {
     assert.deepStrictEqual(convertSdkNameToReportName('eyes-selenium'), 'js_selenium_4')
-    assert.deepStrictEqual(convertSdkNameToReportName('eyes.selenium'), 'js_selenium_3')
+    assert.deepStrictEqual(convertSdkNameToReportName('eyes-selenium-3'), 'js_selenium_3')
     assert.deepStrictEqual(convertSdkNameToReportName('eyes.webdriverio.javascript5'), 'js_wdio_5')
     assert.deepStrictEqual(convertSdkNameToReportName('eyes.webdriverio.javascript4'), 'js_wdio_4')
     assert.deepStrictEqual(convertSdkNameToReportName('eyes-images'), 'js_images')